+++ abstractions/ruby/microsoft_kiota_abstractions/lib/microsoft_kiota_abstractions/request_adapter.rb
@@ -5,9 +5,11 @@ module MicrosoftKiotaAbstractions
   module RequestAdapter
     include ResponseHandler
 
-    def send_async(request_info, type, response_handler)
+    def send_async(request_info, factory, response_handler)
       raise NotImplementedError.new
     end
+
+    # TODO we're most likley missing something for enums and collections or at least at the implemenation level
 
     def get_serialization_writer_factory()
       raise NotImplementedError.new